Modularize info components and adjust responsive
