+++ src/components/Abilities/Abilities.style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const CaptalizeInfo = styled.span`
+  text-transform: capitalize;
+`
+
+export const Topic = styled.div`
+  padding-bottom: 24px;
+  display: grid;
+  grid-template-columns: 0.3fr 1fr;
+  align-items: center;
+`
