+++ src/components/Abilities/index.js
@@ -1,0 +1,3 @@
+import Abilities from './Abilities'
+
+export default Abilities
