+++ src/components/Misc/index.js
@@ -1,0 +1,3 @@
+import Misc from './Misc'
+
+export default Misc
